// src/app/classes/interfaces.ts
export interface RespuestaUser {
  ok: boolean;
  total: number;
  pagina: number;
  uadm: Usuario[];
}
// ...
export interface Usuario {
  avatar?: string;
  _id?: string;
  nombre?: string;
  email?: string;
  password?: string;
  password_c?: string;
  role?: string;
  fb?: string;
  fecha?: string;

  interests?: string;
  version?: string;
  notification_id?: string;
}